Quell intermittent PrimerDesigner errors when run under mod_fcgid instead of mod_perl.  Makes sense to me!

+++ conf/plugins/PrimerDesigner.pm
@@ -102,8 +102,9 @@ use vars qw/@ISA $PNG $CALL/;
 
 # Arg, modperl
 END {
-	
-  CGI::Delete_all();
+  if ($ENV{'MOD_PERL'}) {
+    CGI::Delete_all();
+  }
 }
